First set of functions
Added a function to add a line of data to a csv file. And the fileCheck function checks if the file is empty, if so it prepares it with correct headers. If the file isn't empty it will check to see if this file has the correct headers for the data we want to put in and will let us know if it matches. Not sure if it's really useful.

Also this is working with lists right now, may switch the dictionaries and may look into creating a class for all this. But just want some functionality for now.